+++ classes/Entities/Collection.php
@@ -3,6 +3,10 @@ namespace Grav\Plugin\FlexDirectory\Entities;
 
 use Grav\Plugin\FlexDirectory\SimpleDataContainer;
 
+/**
+ * Class Collection
+ * @package Grav\Plugin\FlexDirectory\Entities
+ */
 class Collection extends SimpleDataContainer
 {
     public function saveDataItem($id, $item)