+++ classes/SimpleCollection.php
@@ -3,7 +3,10 @@ namespace Grav\Plugin\FlexDirectory;
 
 use Grav\Common\Iterator;
 
+/**
+ * Class SimpleCollection
+ * @package Grav\Plugin\FlexDirectory
+ */
 class SimpleCollection extends Iterator
 {
-
 }