+++ classes/FlexCollection.php
@@ -1,8 +1,13 @@
 <?php
 namespace Grav\Plugin\FlexDirectory;
 
+use Grav\Common\Debugger;
+use Grav\Common\Grav;
+use Grav\Common\Twig\Twig;
+use Grav\Framework\ContentBlock\HtmlBlock;
 use Grav\Framework\Object\ObjectCollection;
 use Grav\Plugin\FlexDirectory\Interfaces\FlexCollectionInterface;
+use RocketTheme\Toolbox\Event\Event;
 
 /**
  * Class FlexCollection
@@ -28,12 +33,51 @@ class FlexCollection extends ObjectCollection implements FlexCollectionInterface
     }
 
     /**
+     * @param string $layout
+     * @param array $context
+     * @return HtmlBlock
+     * @throws \Exception
+     * @throws \Throwable
+     * @throws \Twig_Error_Loader
+     * @throws \Twig_Error_Syntax
+     */
+    public function render($layout = 'default', array $context = [])
+    {
+        $grav = Grav::instance();
+
+        /** @var Debugger $debugger */
+        $debugger = $grav['debugger'];
+        $debugger->startTimer('flex-collection-' . $this->getType(), 'Collection ' . $this->getType());
+
+        $block = new HtmlBlock();
+
+        $grav->fireEvent('onFlexCollectiontRender', new Event([
+            'collection' => $this,
+            'layout' => &$layout,
+            'context' => &$context
+        ]));
+
+        $output = $this->getTemplate($layout)->render(
+            ['grav' => $grav, 'block' => $block, 'collection' => $this] + $context
+        );
+
+        $block->setContent($output);
+
+        $debugger->stopTimer('flex-collection-' . $this->getType());
+
+        return $block;
+    }
+
+    /**
      * @param FlexType $type
      * @return $this
      */
     public function setFlexType(FlexType $type)
     {
         $this->flexType = $type;
+
+        // TODO: need a better way
+        static::$type = $type->getType();
 
         return $this;
     }
@@ -62,5 +106,25 @@ class FlexCollection extends ObjectCollection implements FlexCollectionInterface
         }
 
         return $elements;
+    }
+
+    /**
+     * @param string $layout
+     * @return \Twig_Template
+     * @throws \Twig_Error_Loader
+     * @throws \Twig_Error_Syntax
+     */
+    protected function getTemplate($layout)
+    {
+        $grav = Grav::instance();
+
+        /** @var Twig $twig */
+        $twig = $grav['twig'];
+
+        try {
+            return $twig->twig()->resolveTemplate(["flex-directory/layouts/{$this->getType()}/collection/{$layout}.html.twig"]);
+        } catch (\Twig_Error_Loader $e) {
+            return $twig->twig()->resolveTemplate(["flex-directory/layouts/404.html.twig"]);
+        }
     }
 }