+++ classes/Controllers/SiteController.php
@@ -1,0 +1,10 @@
+<?php
+namespace Grav\Plugin\FlexDirectory\Controllers;
+
+/**
+ * Class SiteController
+ * @package Grav\Plugin\FlexDirectory
+ */
+class SiteController extends SimpleController
+{
+}