+++ classes/SimpleCollection.php
@@ -1,9 +1,0 @@
-<?php
-namespace Grav\Plugin\FlexDirectory;
-
-use Grav\Common\Iterator;
-
-class SimpleCollection extends Iterator
-{
-
-}